backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LangChain & LangGraph (aligned)
langchain~=0.2.17 
langchain-community~=0.2.17 
langchain-openai~=0.1.6
langgraph~=0.2.3
langsmith~=0.1.112

# Vector store (Windows-friendly)
chromadb==0.5.23
hnswlib~=0.8.0

# Embeddings & tokenization
sentence-transformers==2.3.1
tiktoken>=0.7.0

# Guardrails
guardrails-ai==0.6.7

# LLM providers
openai>=1.30.0
anthropic>=0.25.0

# Utilities (Py 3.12 safe)
numpy>=1.26,<3
pandas>=2.1,<3
python-multipart==0.0.6
aiofiles==23.2.1
httpx>=0.27.2
anyio>=4.0.0

# Logging & Monitoring
structlog==24.1.0
rich==13.7.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Dev tools
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.12.0

# LLM Providers 
# openai~=1.9.0

# Vector Store & Embeddings
faiss-cpu~=1.7.4
sentence-transformers~=2.3.1
tiktoken